--- plugins/previewer-color/index.html
@@ -1,92 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-
-	<meta charset="utf-8" />
-	<link rel="shortcut icon" href="favicon.png" />
-	<title>Previewer: Color ▲ Prism plugins</title>
-	<base href="../.." />
-	<link rel="stylesheet" href="style.css" />
-	<link rel="stylesheet" href="themes/prism.css" data-noprefix />
-	<link rel="stylesheet" href="plugins/previewer-base/prism-previewer-base.css" data-noprefix />
-	<link rel="stylesheet" href="plugins/previewer-color/prism-previewer-color.css" data-noprefix />
-	<script src="prefixfree.min.js"></script>
-
-	<script>var _gaq = [['_setAccount', 'UA-33746269-1'], ['_trackPageview']];</script>
-	<script src="http://www.google-analytics.com/ga.js" async></script>
-</head>
-<body>
-
-<header>
-	<div class="intro" data-src="templates/header-plugins.html" data-type="text/html"></div>
-
-	<h2>Previewer: Color</h2>
-	<p>Previewer for CSS colors.</p>
-</header>
-
-<section class="language-markup">
-	<h1>How to use</h1>
-
-	<p>You don't need to do anything. With this plugin loaded, a previewer will appear on hovering the color values in code blocks.</p>
-	<p>This plugin is compatible with CSS, Markup attributes, Less, Sass, Scss and Stylus.</p>
-</section>
-
-<section>
-	<h1>Examples</h1>
-
-	<h2>CSS</h2>
-	<pre class="language-css"><code>div {
-	color: rgba(255, 0, 0, 0.2);
-	background: purple;
-	border: 1px solid hsl(100,70%,40%);
-}</code></pre>
-
-	<h2>Markup attributes</h2>
-	<pre class="language-markup"><code>&lt;table bgcolor="#6E5494">
-&lt;tr style="background: lightblue;"></code></pre>
-
-	<h2>Less</h2>
-	<pre class="language-less"><code>@nice-blue: #5B83AD;
-#header a {
-	color: hsla(102, 53%, 42%, 0.4);
-}</code></pre>
-
-	<h2>Sass</h2>
-	<pre class="language-sass"><code>$color: blue
-@mixin foobar
-    color: rgba(147, 32, 34, 0.8)
-.foo
-    color: pink
-</code></pre>
-
-	<h2>Scss</h2>
-	<pre class="language-scss"><code>$color: blue;
-$attr: background;
-.foo {
-    #{$attr}-color: rgba(255,255,0,0.75);
-}</code></pre>
-
-	<h2>Stylus</h2>
-	<pre class="language-stylus"><code>color = olive
-.foo
-	color: #000
-</code></pre>
-
-</section>
-
-<footer data-src="templates/footer.html" data-type="text/html"></footer>
-
-<script src="prism.js"></script>
-<script src="components/prism-less.js"></script>
-<script src="components/prism-sass.js"></script>
-<script src="components/prism-scss.js"></script>
-<script src="components/prism-stylus.js"></script>
-<script src="plugins/previewer-base/prism-previewer-base.js"></script>
-<script src="plugins/previewer-color/prism-previewer-color.js"></script>
-<script src="utopia.js"></script>
-<script src="components.js"></script>
-<script src="code.js"></script>
-
-
-</body>
-</html>
