--- plugins/previewer-easing/prism-previewer-easing.css
@@ -1,29 +1,0 @@
-.prism-previewer-easing {
-	margin-top: -76px;
-	margin-left: -30px;
-	width: 60px;
-	height: 60px;
-	background: #333;
-}
-.prism-previewer-easing.flipped {
-	margin-bottom: -116px;
-}
-.prism-previewer-easing svg {
-	width: 60px;
-	height: 60px;
-}
-.prism-previewer-easing circle {
-	fill: hsl(200, 10%, 20%);
-	stroke: white;
-}
-.prism-previewer-easing path {
-	fill: none;
-	stroke: white;
-	stroke-linecap: round;
-	stroke-width: 4;
-}
-.prism-previewer-easing line {
-	stroke: white;
-	stroke-opacity: 0.5;
-	stroke-width: 2;
-}
